requirements.txt
aiogram==3.7.0
apscheduler==3.10.4
aiosqlite==0.20.0
python-dotenv==1.0.1
pytz==2024.1
